requirements.txt
flask
python-dotenv
google
google-genai
aiohttp
requests